+++ CombinationsConfig.cs
@@ -1,0 +1,20 @@
+﻿using System.ComponentModel;
+using Terraria.ModLoader.Config;
+
+namespace Combinations
+{
+    public class CombinationsConfig : ModConfig
+    {
+        public override ConfigScope Mode => ConfigScope.ServerSide;
+
+        // Automatically set by tModLoader
+        public static CombinationsConfig Instance;
+
+        [Header("Visual")]
+
+        [Label("Hide Molten Shield Inferno")]
+        [Tooltip("Hide Ring of Inferno effect of Molten Shield")]
+        [DefaultValue(false)]
+        public bool HideMoltenShieldRing;
+    }
+}